rename DbgPrintPtrH.h to DbgPrintHelper.h
