+++ core/modules/query/FromList.cc
@@ -35,7 +35,7 @@
 #include <algorithm>
 #include <iterator>
 
-#include "../util/DbgPrintPtrH.h"
+#include "util/DbgPrintHelper.h"
 // Third-party headers
