+++ core/modules/query/HavingClause.cc
@@ -40,7 +40,7 @@
 #include "query/BoolTerm.h"
 #include "query/QueryTemplate.h"
 #include "util/PointerCompare.h"
-#include "util/DbgPrintPtrH.h"
+#include "util/DbgPrintHelper.h"
 
 namespace lsst {
 namespace qserv {
